Add files via upload
Divided script into functions and main program
